resultats/repositories/equipe_09_alterego/idfm_hackaton_2024-main/backend/api-llm/src/pyproject.toml
[tool.poetry]
package-mode = false
name = "api-llm_2024"
version = "0.1.0"
description = ""

[tool.poetry.dependencies]
python = "^3.12"
sqlfluff = "^3.0.6"
pandas = "^2.2.2"
openpyxl = "^3.1.4"
duckdb = "^1.1.3"
polars = "^1.14.0"
networkx = "^3.4.2"
plotly = "^5.24.1"
seaborn = "^0.13.2"
pyarrow = "^18.0.0"
fastapi = {extras = ["standard"], version = "^0.115.5"}
uvicorn = "^0.32.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
openai = "^1.55.0"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.3.1"
jupytext = "^1.16.4"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
